// apps/api/src/users/enums/user-role.enum.ts
export enum UserRole {
  PRINCIPAL = 'principal',
  DEPUTY_PRINCIPAL = 'deputy_principal',
  HOD = 'hod',
  CLASS_TEACHER = 'class_teacher',
  ACCOUNTANT = 'accountant',
  PARENT = 'parent',
  NURSE = 'nurse',
}

export enum Form {
  FORM_1 = 'form_1',
  FORM_2 = 'form_2',
  FORM_3 = 'form_3',
  FORM_4 = 'form_4',
}

export enum Stream {
  A = 'A',
  B = 'B',
  C = 'C',
  D = 'D',
}

export enum Gender {
  MALE = 'male',
  FEMALE = 'female',
}